move wolfram over to new pattern, getFull doesn't work so it is commented out
