--- features/wolfram.js
@@ -1,42 +1,0 @@
-require('dotenv').config()
-var preferences = require('./userPreferences');
-
-const WolframAlphaAPI = require('wolfram-alpha-api');
-const waApi = WolframAlphaAPI(process.env.WOLFRAM_APPID);
-
-const addLinkToResult = queryString => { 
-  return `https://www.wolframalpha.com/input/?i=${queryString}`;
-}
-
-const wolframGetShort = function(message) {
-  const queryString = message.content.substring(3);
-
-  const userPreferences = preferences.getPreference(`${message.author.username}#${message.author.discriminator}`);
-  // console.log('got:', userPreferences);
-
-  const result = waApi.getShort(  {
-    i: queryString,
-    units: userPreferences.units,
-  });
-
-  return result;
-};
-
-const wolframGetFull = function (message) {
-  const queryString = message.content.substring(3);
-
-  const userPreferences = preferences.getPreference(`${message.author.username}#${message.author.discriminator}`);
-  // console.log('got:', userPreferences);
-
-  const result = waApi.getFull({
-    i: queryString,
-    units: userPreferences.units,
-  });
-
-  return result;
-};
-
-module.exports = {
-  wolframGetShort,
-  wolframGetFull,
-};
